Add deprecation module to centralize deprecation messages and bits of code
